+++ ouroboros-network/src/Ouroboros/Network/Diffusion/Configuration.hs
@@ -7,14 +7,12 @@ module Ouroboros.Network.Diffusion.Configuration
   ( defaultAcceptedConnectionsLimit
   , defaultPeerSharing
   , defaultBlockFetchConfiguration
-  , defaultChainSyncTimeout
   , defaultDeadlineTargets
   , defaultDeadlineChurnInterval
   , defaultBulkChurnInterval
     -- re-exports
   , AcceptedConnectionsLimit (..)
   , BlockFetchConfiguration (..)
-  , ChainSyncTimeout (..)
   , DiffusionMode (..)
   , MiniProtocolParameters (..)
   , PeerSelectionTargets (..)
@@ -37,7 +35,6 @@ module Ouroboros.Network.Diffusion.Configuration
   ) where
 
 import Control.Monad.Class.MonadTime.SI
-import System.Random (randomRIO)
 
 import Cardano.Network.ConsensusMode
 import Ouroboros.Network.BlockFetch (BlockFetchConfiguration (..),
@@ -45,8 +42,7 @@ import Ouroboros.Network.BlockFetch (BlockFetchConfiguration (..),
 import Ouroboros.Network.ConnectionManager.Core (defaultProtocolIdleTimeout,
            defaultResetTimeout, defaultTimeWaitTimeout)
 import Ouroboros.Network.Diffusion.Policies (closeConnectionTimeout,
-           deactivateTimeout, maxChainSyncTimeout, minChainSyncTimeout,
-           peerMetricsConfiguration)
+           deactivateTimeout, peerMetricsConfiguration)
 import Ouroboros.Network.NodeToNode (DiffusionMode (..),
            MiniProtocolParameters (..), defaultMiniProtocolParameters)
 import Ouroboros.Network.PeerSelection.Governor.Types
@@ -54,9 +50,7 @@ import Ouroboros.Network.PeerSelection.Governor.Types
 import Ouroboros.Network.PeerSelection.PeerSharing (PeerSharing (..))
 import Ouroboros.Network.PeerSharing (ps_POLICY_PEER_SHARE_MAX_PEERS,
            ps_POLICY_PEER_SHARE_STICKY_TIME)
-import Ouroboros.Network.Protocol.ChainSync.Codec (ChainSyncTimeout (..))
 import Ouroboros.Network.Protocol.Handshake (handshake_QUERY_SHUTDOWN_DELAY)
-import Ouroboros.Network.Protocol.Limits (shortWait)
 import Ouroboros.Network.Server.RateLimiting (AcceptedConnectionsLimit (..))
 
 -- | Outbound governor targets
@@ -103,33 +97,6 @@ defaultBlockFetchConfiguration bfcSalt =
     bfcGenesisBFConfig        = GenesisBlockFetchConfiguration
       { gbfcGracePeriod = 10 },  -- seconds
     bfcSalt }
-
-defaultChainSyncTimeout :: IO ChainSyncTimeout
-defaultChainSyncTimeout = do
-    -- These values approximately correspond to false positive
-    -- thresholds for streaks of empty slots with 99% probability,
-    -- 99.9% probability up to 99.999% probability.
-    -- t = T_s [log (1-Y) / log (1-f)]
-    -- Y = [0.99, 0.999...]
-    -- T_s = slot length of 1s.
-    -- f = 0.05
-    -- The timeout is randomly picked per bearer to avoid all bearers
-    -- going down at the same time in case of a long streak of empty
-    -- slots.
-    -- To avoid global synchronosation the timeout is picked uniformly
-    -- from the interval 135 - 269, corresponds to the a 99.9% to
-    -- 99.9999% thresholds.
-    -- TODO: The timeout should be drawn at random everytime chainsync
-    --       enters the must reply state. A static per connection timeout
-    --       leads to selection preassure for connections with a large
-    --       timeout, see #4244.
-    mustReplyTimeout <- Just . realToFrac <$> randomRIO ( realToFrac minChainSyncTimeout :: Double
-                                                        , realToFrac maxChainSyncTimeout :: Double
-                                                        )
-    return ChainSyncTimeout { canAwaitTimeout  = shortWait,
-                              intersectTimeout = shortWait,
-                              mustReplyTimeout,
-                              idleTimeout      = Just 3673 }
 
 defaultDeadlineChurnInterval :: DiffTime
 defaultDeadlineChurnInterval = 3300