+++ ouroboros-network/testlib/Test/Ouroboros/Network/Diffusion/Testnet/Cardano.hs
@@ -349,7 +349,6 @@ unit_cm_valid_transitions =
             [ ( NodeArgs
                   (-2)
                   InitiatorAndResponderDiffusionMode
-                  (Just 269)
                   (Map.fromList [(RelayAccessAddress "0:71:0:1:0:1:0:1" 65_534,
                                   DoAdvertisePeer)])
                   GenesisMode
@@ -392,7 +391,6 @@ unit_cm_valid_transitions =
               , ( NodeArgs
                   0
                   InitiatorAndResponderDiffusionMode
-                  (Just 90)
                   Map.empty
                   GenesisMode
                   (Script (DontUseBootstrapPeers :| []))
@@ -568,7 +566,6 @@ unit_connection_manager_trace_coverage =
           (NodeArgs {
              naSeed = 0,
              naDiffusionMode = InitiatorAndResponderDiffusionMode,
-             naMbTime = Just 224,
              naPublicRoots = Map.empty,
              naConsensusMode = PraosMode,
              naBootstrapPeers = Script (DontUseBootstrapPeers :| []),
@@ -598,7 +595,6 @@ unit_connection_manager_trace_coverage =
           (NodeArgs {
              naSeed = 0,
              naDiffusionMode = InitiatorAndResponderDiffusionMode,
-             naMbTime = Just 224,
              naPublicRoots = Map.empty,
              naConsensusMode = PraosMode,
              naBootstrapPeers = Script (DontUseBootstrapPeers :| []),
@@ -692,7 +688,6 @@ unit_connection_manager_transitions_coverage =
           (NodeArgs {
              naSeed = 0,
              naDiffusionMode = InitiatorAndResponderDiffusionMode,
-             naMbTime = Just 224,
              naPublicRoots = Map.empty,
              naConsensusMode = PraosMode,
              naBootstrapPeers = Script (DontUseBootstrapPeers :| []),
@@ -722,7 +717,6 @@ unit_connection_manager_transitions_coverage =
           (NodeArgs {
              naSeed = 0,
              naDiffusionMode = InitiatorAndResponderDiffusionMode,
-             naMbTime = Just 224,
              naPublicRoots = Map.empty,
              naConsensusMode = PraosMode,
              naBootstrapPeers = Script (DontUseBootstrapPeers :| []),
@@ -1014,7 +1008,7 @@ unit_4177 = prop_inbound_governor_transitions_coverage absNoAttenuation script
     script =
       DiffusionScript (SimArgs 1 10)
         (singletonTimedScript Map.empty)
-        [ ( NodeArgs (-6) InitiatorAndResponderDiffusionMode (Just 180)
+        [ ( NodeArgs (-6) InitiatorAndResponderDiffusionMode
               (Map.fromList [(RelayAccessDomain "test2" 65_535, DoAdvertisePeer)])
               PraosMode
               (Script (UseBootstrapPeers [RelayAccessDomain "bootstrap" 0] :| []))
@@ -1045,7 +1039,7 @@ unit_4177 = prop_inbound_governor_transitions_coverage absNoAttenuation script
             ,Reconfigure 4.870_967_741_935 [(1,1,Map.fromList [(RelayAccessDomain "test2" 65_535,LocalRootConfig DoAdvertisePeer InitiatorAndResponderDiffusionMode IsNotTrustable)])]
             ]
           )
-        , ( NodeArgs 1 InitiatorAndResponderDiffusionMode (Just 135)
+        , ( NodeArgs 1 InitiatorAndResponderDiffusionMode
              (Map.fromList [(RelayAccessAddress "0:7:0:7::" 65_533, DoAdvertisePeer)])
              PraosMode
               (Script (UseBootstrapPeers [RelayAccessDomain "bootstrap" 0] :| []))
@@ -1670,7 +1664,6 @@ unit_4191 = testWithIOSim prop_diffusion_dns_can_recover long_trace absInfo scri
         [(NodeArgs
             16
             InitiatorAndResponderDiffusionMode
-            (Just 224)
             Map.empty
             PraosMode
             (Script (UseBootstrapPeers [RelayAccessDomain "bootstrap" 0] :| []))
@@ -1786,7 +1779,6 @@ prop_connect_failure (AbsIOError ioerr) =
         [ (NodeArgs {
             naSeed = 0,
             naDiffusionMode = InitiatorAndResponderDiffusionMode,
-            naMbTime = Just 224,
             naPublicRoots = Map.empty,
             naConsensusMode = PraosMode,
             naBootstrapPeers = Script (DontUseBootstrapPeers :| []),
@@ -1815,7 +1807,6 @@ prop_connect_failure (AbsIOError ioerr) =
           (NodeArgs {
             naSeed = 0,
             naDiffusionMode = InitiatorAndResponderDiffusionMode,
-            naMbTime = Just 224,
             naPublicRoots = Map.empty,
             naConsensusMode = PraosMode,
             naBootstrapPeers = Script (DontUseBootstrapPeers :| []),
@@ -1912,7 +1903,6 @@ prop_accept_failure (AbsIOError ioerr) =
         [ (NodeArgs {
             naSeed = 0,
             naDiffusionMode = InitiatorAndResponderDiffusionMode,
-            naMbTime = Just 224,
             naPublicRoots = Map.empty,
             naConsensusMode = PraosMode,
             naBootstrapPeers = Script (DontUseBootstrapPeers :| []),
@@ -1941,7 +1931,6 @@ prop_accept_failure (AbsIOError ioerr) =
           (NodeArgs {
             naSeed = 0,
             naDiffusionMode = InitiatorAndResponderDiffusionMode,
-            naMbTime = Just 224,
             naPublicRoots = Map.empty,
             naConsensusMode = PraosMode,
             naBootstrapPeers = Script (DontUseBootstrapPeers :| []),
@@ -2968,7 +2957,6 @@ async_demotion_network_script =
     common = NodeArgs {
         naSeed             = 10,
         naDiffusionMode    = InitiatorAndResponderDiffusionMode,
-        naMbTime           = Just 1,
         naPublicRoots      = Map.empty,
         naConsensusMode    = PraosMode,
         naBootstrapPeers   = Script (UseBootstrapPeers [RelayAccessDomain "bootstrap" 0] :| []),
@@ -3547,8 +3535,8 @@ prop_unit_4258 =
       diffScript = DiffusionScript
         (SimArgs 1 10)
         (singletonTimedScript Map.empty)
-        [( NodeArgs (-3) InitiatorAndResponderDiffusionMode (Just 224)
-             Map.empty
+        [( NodeArgs (-3) InitiatorAndResponderDiffusionMode
+             (Map.fromList [])
              PraosMode
              (Script (UseBootstrapPeers [RelayAccessDomain "bootstrap" 0] :| []))
              (TestAddress (IPAddr (read "0.0.0.4") 9))
@@ -3581,7 +3569,7 @@ prop_unit_4258 =
              Reconfigure 4.190_476_190_476 []
            ]
          ),
-         ( NodeArgs (-5) InitiatorAndResponderDiffusionMode (Just 269)
+         ( NodeArgs (-5) InitiatorAndResponderDiffusionMode
              (Map.fromList [(RelayAccessAddress "0.0.0.4" 9, DoAdvertisePeer)])
              PraosMode
              (Script (UseBootstrapPeers [RelayAccessDomain "bootstrap" 0] :| []))
@@ -3653,7 +3641,6 @@ prop_unit_reconnect =
           [(NodeArgs
               (-3)
               InitiatorAndResponderDiffusionMode
-              (Just 224)
               Map.empty
               PraosMode
               (Script (DontUseBootstrapPeers :| []))
@@ -3684,7 +3671,6 @@ prop_unit_reconnect =
           , (NodeArgs
                (-1)
                InitiatorAndResponderDiffusionMode
-               (Just 2)
                Map.empty
                PraosMode
                (Script (DontUseBootstrapPeers :| []))
@@ -4109,7 +4095,6 @@ unit_peer_sharing =
     defaultNodeArgs naConsensusMode = NodeArgs {
         naSeed = 0,
         naDiffusionMode = InitiatorAndResponderDiffusionMode,
-        naMbTime = Nothing,
         naPublicRoots = mempty,
         naBootstrapPeers = singletonScript DontUseBootstrapPeers,
         naAddr = undefined,
@@ -4615,7 +4600,6 @@ unit_local_root_diffusion_mode diffusionMode =
           (NodeArgs {
              naSeed = 0,
              naDiffusionMode = InitiatorAndResponderDiffusionMode,
-             naMbTime = Just 224,
              naPublicRoots = Map.empty,
              naConsensusMode = PraosMode,
              naBootstrapPeers = Script (DontUseBootstrapPeers :| []),
@@ -4645,7 +4629,6 @@ unit_local_root_diffusion_mode diffusionMode =
           (NodeArgs {
              naSeed = 0,
              naDiffusionMode = InitiatorAndResponderDiffusionMode,
-             naMbTime = Just 224,
              naPublicRoots = Map.empty,
              naConsensusMode = PraosMode,
              naBootstrapPeers = Script (DontUseBootstrapPeers :| []),